refactor(k3): whale policy and startup dance move next to the data plane
k3_whale_width/admits/gang/segments and their tests move from
scheduler/whale.rs to executor/cp.rs, beside the CP admission and split
they parameterize — the rendezvous module keeps only the protocol state
machines. The executor's whale startup sequence (arm slab, import world,
install gang, re-arm scratch) moves from executor/mod.rs into
whale_gang.rs impl blocks, so the fleet data plane is one file.
